license-plate-detection/requirements.txt
numpy
opencv-python-headless
ffmpeg-python
pylabel
pytesseract
python-Levenshtein